+++ CHANGELOG.md
@@ -3,5 +3,5 @@ All notable changes to this project will be documented in this file.
 
 ## [0.0.2]
 ### Added
-- Unlikelihood candidates loss (`--unlikelihood_alpha` added as a 
-`train_model.py` cli argument).
+- Added telegram client for text generator service.
+- Added unlikelihood candidates loss.